# docs/openapi.yaml
openapi: 3.0.3
info:
  title: Nexora API
  version: 1.0.0
  description: 'Nexora is a multi-currency digital asset platform built on Stellar. It provides

    wallets, instant transfers, FX conversions, scheduled recurring payments,

    batch disbursements, fiat on/off-ramps, and a developer-friendly webhook system.


    All endpoints under `/v1` require a Bearer API key in the `Authorization` header.


    ## API key scopes


    An API key can be limited to specific resources and actions. Scopes have the

    form `<resource>:<action>`, where action is `read` or `write`:


    | Resource | Scopes | Covers |

    |---|---|---|

    | wallets | `wallets:read`, `wallets:write` | `/v1/wallets` (create, trustlines, contract wallet changes) |

    | transfers | `transfers:read`, `transfers:write` | `/v1/transfers`, `/v1/transactions`, `/v1/schedules`, `/v1/refunds` |

    | batches | `batches:read`, `batches:write` | `/v1/transfers/batch` |

    | webhooks | `webhooks:read`, `webhooks:write` | `/v1/webhooks` |

    | reports | `reports:read`, `reports:write` | `/v1/usage` (`reports:write` is reserved for report generation) |


    Other scopes: `keys:*`, `audit:read`, `fiat:*`, `fx:*`, `fees:read`,

    `beneficiaries:*`, `compliance:*`. `<resource>:*` grants both actions on one

    resource and `*` grants everything.


    On scoped resource groups, `GET`/`HEAD` need the `read` scope and every other

...

...

...

    path only; never the key itself).


...

...

...

...

...

...

    '
  contact:
...
servers:
...
security:
- BearerAuth: []
paths:
...
  /v1/transfers/batch:
    post:
      operationId: createBatch
      summary: Create a batch transfer
      description: 'Submits up to 100 transfers from a single source wallet in one request.

        Each transfer is processed independently; partial success is possible.

        '
      tags:
      - Transfers
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateBatchRequest'
            example:
              from_wallet_id: w_1a2b3c4d
              transfers:
              - to_wallet_id: w_5e6f7g8h
                asset: USDC
                amount: '50.0000000'
                reference: payroll-jan
              - to_wallet_id: w_9i0j1k2l
                asset: USDC
                amount: '75.0000000'
                reference: payroll-jan
      parameters:
      - name: Idempotency-Key
        in: header
        required: true
        description: 'Client-generated UUID v4 that makes this request safe to retry.

          The result is held for 24 hours: a repeat with the same body

          replays the original response, a repeat with a different body or

          a concurrent repeat is rejected with 409.

          '
        schema:
          type: string
          format: uuid
      responses:
        '202':
          description: Batch accepted
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BatchResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '409':
          $ref: '#/components/responses/Conflict'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /v1/transfers/batch/{batchId}:
    get:
      operationId: getBatch
      summary: Get batch status
      description: Returns the batch and the status of every individual transfer in it.
      tags:
      - Transfers
      parameters:
      - name: batchId
        in: path
        required: true
        schema:
          type: string
        description: Batch ID
      responses:
        '200':
          description: Batch details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BatchResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /v1/transfers/batch/{batchId}/export:
    get:
      operationId: exportBatch
      summary: Export batch as CSV
      description: Downloads the batch transfer results as a CSV file.
      tags:
      - Transfers
      parameters:
      - name: batchId
        in: path
        required: true
        schema:
          type: string
        description: Batch ID
      responses:
        '200':
          description: CSV file
          content:
            text/csv:
              schema:
                type: string
                format: binary
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
...
  /v1/audit:
    get:
      operationId: listAuditEvents
      summary: List audit events
      description: Returns append-only tenant audit events with filtering by action, actor, resource, and time range.
      tags:
      - Audit
      parameters:
      - name: action
        in: query
        required: false
        schema:
          type: string
        description: Filter by action name
      - name: actor
        in: query
        required: false
        schema:
          type: string
        description: Filter by actor ID
      - name: resource_type
        in: query
        required: false
        schema:
          type: string
        description: Filter by resource type
      - name: resource_id
        in: query
        required: false
        schema:
          type: string
        description: Filter by resource ID
      - name: from
        in: query
        required: false
        schema:
          type: string
          format: date-time
        description: Start of time range
      - name: to
        in: query
        required: false
        schema:
          type: string
          format: date-time
        description: End of time range
      - name: limit
        in: query
        required: false
        schema:
          type: integer
          default: 50
        description: Maximum results
      - name: offset
        in: query
        required: false
        schema:
          type: integer
          default: 0
        description: Offset for pagination
      responses:
        '200':
          description: List of audit events
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuditListResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /v1/audit/export:
    get:
      operationId: exportAuditEvents
      summary: Export audit events
      description: Exports tenant audit events as CSV or JSON.
      tags:
      - Audit
      parameters:
      - name: format
        in: query
        required: false
        schema:
          type: string
          enum:
          - csv
          - json
          default: json
        description: Export format
      - name: action
        in: query
        required: false
        schema:
          type: string
        description: Filter by action name
      - name: actor
        in: query
        required: false
        schema:
          type: string
        description: Filter by actor ID
      - name: resource_type
        in: query
        required: false
        schema:
          type: string
        description: Filter by resource type
      - name: resource_id
        in: query
        required: false
        schema:
          type: string
        description: Filter by resource ID
      - name: from
        in: query
        required: false
        schema:
          type: string
          format: date-time
        description: Start of time range
      - name: to
        in: query
        required: false
        schema:
          type: string
          format: date-time
        description: End of time range
      responses:
        '200':
          description: Exported audit events
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/AuditEvent'
            text/csv:
              schema:
                type: string
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '500':
          $ref: '#/components/responses/InternalServerError'
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: token
      description: 'Pass a tenant API key (`sk_live_...` or `sk_test_...`) or JWT bearer token.

        In Swagger UI, enter your token in the Authorize dialog (the ''Bearer '' prefix is added automatically).

        '
    ApiKeyAuth:
      type: apiKey
      in: header
      name: Authorization
      description: 'Pass an API key as `Bearer <key>` in the `Authorization` header.

        '
    AdminAuth:
      type: http
      scheme: bearer
      bearerFormat: admin token
      description: Administrative bearer token.
...
  responses:
    BadRequest:
      description: Bad request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error:
              code: BAD_REQUEST
              message: invalid request body
              status: 400
              request_id: req-example
    Unauthorized:
      description: Unauthorized — missing or invalid API key
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error:
              code: UNAUTHORIZED
              message: missing or invalid authorization header
              status: 401
              request_id: req-example
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error:
              code: NOT_FOUND
              message: resource not found
              status: 404
              request_id: req-example
    Forbidden:
      description: Forbidden — admin access required
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error:
              code: FORBIDDEN
              message: admin access required
              status: 403
              request_id: req-example
    Conflict:
      description: Conflict — a request with this Idempotency-Key is still in flight
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error:
              code: REQUEST_IN_PROGRESS
              message: a request with this Idempotency-Key is still in progress
              status: 409
              request_id: req-example
...
    InternalServerError:
      description: Internal server error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error:
              code: INTERNAL_ERROR
              message: an unexpected error occurred
              status: 500
              request_id: req-example
  schemas:
...
    ErrorResponse:
      type: object
      required:
      - error
      properties:
        error:
          type: object
          required:
          - code
          - message
          - status
          - request_id
          properties:
            code:
              type: string
              enum:
              - BAD_REQUEST
              - ANCHOR_DEPOSIT_FAILED
              - ANCHOR_REGISTRATION_FAILED
              - ANCHOR_TRANSACTION_NOT_FOUND
              - ANCHOR_WITHDRAWAL_FAILED
              - API_KEY_NOT_FOUND
              - API_KEY_REVOKED
              - API_KEY_EXPIRED
              - BENEFICIARY_EXISTS
              - BENEFICIARY_NOT_ACTIVE
              - BENEFICIARY_NOT_ALLOWED
              - CONFIG_NOT_FOUND
              - CONFLICT
              - COOLING_OFF
              - CROSS_MODE_KEY_OPERATION
              - ENVIRONMENT_REQUIRED
              - IDEMPOTENCY_KEY_REQUIRED
              - IDEMPOTENCY_RECORD_PERSISTENCE_FAILED
              - IDEMPOTENCY_RESPONSE_TOO_LARGE
              - INSUFFICIENT_SCOPE
              - INVALID_ENVIRONMENT_MODE
              - INVALID_PREDICATE
              - INVALID_REQUEST
              - LAST_ORG_OWNER
              - METHOD_NOT_ALLOWED
              - MISSING_TRUSTLINE
              - NOT_IMPLEMENTED
              - PAYMENT_LINK_UNAVAILABLE
              - PREDICATE_NOT_SATISFIABLE
              - RATE_LIMITED
              - REFUND_CONFLICT
              - TEST_MODE_REQUIRED
              - TRANSFER_ID_INVALID
              - TRANSFER_ID_REQUIRED
              - TRANSFER_NOT_CANCELLABLE
              - UNSUPPORTED_FIAT_CURRENCY
              - WALLET_ID_REQUIRED
              - NOT_FOUND
              - UNAUTHORIZED
              - FORBIDDEN
              - INSUFFICIENT_SCOPE
              - INTERNAL_ERROR
              - INSUFFICIENT_SWEEPABLE_BALANCE
              - QUOTE_EXPIRED
              - QUOTE_ALREADY_USED
              - REVIEW_ALREADY_DECIDED
              - CLAIMABLE_BALANCE_NOT_PENDING
              - REQUEST_IN_PROGRESS
              description: Machine-readable error code
            message:
              type: string
              description: Human-readable error message
            status:
              type: integer
              description: HTTP status code; matches the response status
              minimum: 400
              maximum: 599
            request_id:
              type: string
              description: Effective X-Request-ID response header value
...
    AuditEvent:
      type: object
      properties:
        id:
          type: string
          format: uuid
        tenant_id:
          type: string
          format: uuid
        mode:
          type: string
          enum:
          - test
          - live
        actor_type:
          type: string
          enum:
          - api_key
          - user
          - system
        actor_id:
          type: string
        action:
          type: string
        resource_type:
          type: string
        resource_id:
          type: string
        metadata:
          type: object
        ip_address:
          type: string
          nullable: true
        user_agent:
          type: string
          nullable: true
        created_at:
          type: string
          format: date-time
    AuditListResponse:
      type: object
      properties:
        events:
          type: array
          items:
            $ref: '#/components/schemas/AuditEvent'
        total:
          type: integer
        limit:
          type: integer
        offset:
          type: integer
    CreateBatchRequest:
      type: object
      required:
      - from_wallet_id
      - transfers
      properties:
        from_wallet_id:
          type: string
          format: uuid
        transfers:
          type: array
          minItems: 1
          maxItems: 100
          items:
            $ref: '#/components/schemas/BatchItemRequest'
    BatchItemRequest:
      type: object
      required:
      - to_wallet_id
      - asset
      - amount
      properties:
        to_wallet_id:
          type: string
          format: uuid
        asset:
          type: string
        amount:
          type: string
          description: Positive decimal amount
        reference:
          type: string
          description: Optional reference for this transfer
    BatchResponse:
      type: object
      properties:
        id:
          type: string
        status:
          type: string
          enum:
          - pending
          - processing
          - partial
          - completed
          - failed
          - compliance_hold
        total_count:
          type: integer
        success_count:
          type: integer
        failed_count:
          type: integer
        held_count:
          type: integer
        created_at:
          type: string
          format: date-time
        transfers:
          type: array
          items:
            $ref: '#/components/schemas/BatchTransferItem'
    BatchTransferItem:
      type: object
      properties:
        id:
          type: string
        to_wallet_id:
          type: string
        asset:
          type: string
        amount:
          type: string
        reference:
          type: string
          nullable: true
        status:
          type: string
          enum:
          - pending
          - submitted
          - confirmed
          - failed
        tx_hash:
          type: string
          nullable: true
        failure_reason:
          type: string
          description: Machine-readable reason code when the transfer failed during initiation
        failure_message:
          type: string
          description: Underlying initiation error for support diagnostics
...